Improve Error()

+++ examples/myapp/entity.go
@@ -1,6 +1,6 @@
 package myapp
 
 type User struct {
-	ID       int64  `db:"id"`
+	ID       int    `db:"id"`
 	Username string `db:"username"`
 }
